move shcema
