+++ models/contact-model.js
@@ -1,5 +1,4 @@
 const { Schema, model } = require("mongoose");
-const Joi = require("joi");
 
 const { handleMongooseError } = require("../helpers");
 
@@ -27,33 +26,8 @@ const contactSchema = new Schema(
 
 contactSchema.post("save", handleMongooseError);
 
-const addSchema = Joi.object({
-  name: Joi.string().required().messages({
-    "any.required": "missing required name field",
-  }),
-  email: Joi.string().required().messages({
-    "any.required": "missing required email field",
-  }),
-  phone: Joi.string().required().messages({
-    "any.required": "missing required phone field",
-  }),
-  favorite: Joi.boolean(),
-});
-
-const updateFavoriteSchema = Joi.object({
-  favorite: Joi.boolean().required().messages({
-    "any.required": "missing field favorite",
-  }),
-});
-
 const Contact = model("contact", contactSchema);
-
-const schemas = {
-  addSchema,
-  updateFavoriteSchema,
-};
 
 module.exports = {
   Contact,
-  schemas,
 };
